Move default timeouts from request to client configuration.
The current implementation does not allow the configuration of custom timeouts since they are set to 20s/10s for each request. Passing in a custom client with specified timeouts does not change the timeout since they are overridden. In some scenarios, it is desirable to have shorter timeouts to avoid stalling the processing of requests/jobs.

This change should not modify the library's default behaviour but move the timeouts into the client configuration to allow the customization of the values by providing a different instance of the Guzzle client.

Signed-off-by: Claude Vessaz <claude.vessaz@hootsuite.com>